Course Complete

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ public function store(Request $request) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
 //        Session::flash('message', 'Listing Created');
@@ -61,6 +63,12 @@ public function store(Request $request) {
 
 // Update Listing Data
     public function update(Request $request, Listing $listing) {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -84,7 +92,17 @@ public function store(Request $request) {
 
     //Delete Listing
     public function destroy(Listing $listing) {
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfully!');
+    }
+
+    //Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
